+++ app/middlewares/navigation/routes.js
@@ -9,5 +9,5 @@ export const HOME = 'Home'
 export const ONBOARDING_AGREE = 'OnboardingAgree'
 export const ONBOARDING_INITIALIZE = 'OnboardingInitialize'
 export const ONBOARDING_REQUEST = 'OnboardingRequest'
-export const ONBOARDING_SUBMIT_CODE = 'OnboardingSubmitCode'
+export const ONBOARDING_SUBMIT = 'OnboardingSubmit'
 export const SPLASH = 'Splash'
