--- app/screens/OnboardingSubmitCode/index.js
@@ -1,4 +1,0 @@
-import OnboardingSubmitCode from './OnboardingSubmitCode'
-import { connect } from './presenter'
-
-export default connect(OnboardingSubmitCode)
