+++ app/screens/index.js
@@ -9,5 +9,5 @@ export { default as Home } from './Home'
 export { default as OnboardingAgree } from './OnboardingAgree'
 export { default as OnboardingInitialize } from './OnboardingInitialize'
 export { default as OnboardingRequest } from './OnboardingRequest'
-export { default as OnboardingSubmitCode } from './OnboardingSubmitCode'
+export { default as OnboardingSubmit } from './OnboardingSubmit'
 export { default as Splash } from './Splash'
